Parallel processing handling function

Parallel processing handling function

Usage

Boot_parallel(
  cluster,
  seed.cl,
  samples,
  data,
  formula.str,
  group.name,
  groups,
  meta.info,
  a1,
  a2,
  trend,
  robust,
  permutating.group
)

Arguments

cluster

A parallel cluster object for distributed computation, e.g., created by makeCluster(). Default is 2.

seed.cl

An integer specifying the seed for randomization; if not provided, the default is 1234.

samples

bootstrapped samples matrix

data

A SummarizedExperiment object or a matrix where rows represent features (e.g., genes, proteins) and columns represent samples. The values should be log-transformed.

formula.str

A formula string used when covariates are present in meta. info for modeling. It should include "~ 0 + ..." to exclude the intercept from the model.

group.name

A string specifying the column in meta.info that represents the groups or conditions for comparison.

groups

groups information from meta.info

meta.info

A data frame containing sample-level metadata, where each row corresponds to a sample. It should include the grouping variable specified in group.name. If x is a SummarizedExperiment object, meta.info must be a vector of the metadata needed for the model to run and can be retrieved using colData().

a1

Optional numeric value used in the optimization process. If defined by the user, no optimization occurs.

a2

Optional numeric value used in the optimization process. If defined by the user, no optimization occurs.

trend

indicating whether to include trend fitting in the differential expression analysis. Default is TRUE. see eBayes.

robust

indicating whether robust fitting should be used. Default is TRUE, see eBayes.

permutating.group

Logical, If TRUE, the permutation for calculating the null distribution is performed by permuting the target group only specified in group.name. If FALSE, the entire meta.info will be permuted (recommended to be set to FALSE).

Value

A list containing: D, S, pD, pS for bootstrapped data and for permuted data.