Parallel processing handling function

Parallel processing handling function

Boot_parallel(
  BPPARAM = NULL,
  samples,
  data,
  formula.str,
  group.name,
  groups,
  meta.info,
  a1,
  a2,
  pSamples
)

Arguments

BPPARAM

A parallel BPPARAM object for distributed computation.

samples

bootstrapped samples matrix

data

A SummarizedExperiment object or a matrix where rows represent features (e.g., genes, proteins) and columns represent samples. The values should be log-transformed.

formula.str

A formula string used when covariates are present in meta. info for modeling. It should include "~ 0 + ..." to exclude the intercept from the model.

group.name

A string specifying the column in meta.info that represents the groups or conditions for comparison.

groups

groups information from meta.info

meta.info

A data frame containing sample-level metadata, where each row corresponds to a sample. It should include the grouping variable specified in group.name. If x is a SummarizedExperiment object, meta.info must be a vector of the metadata needed for the model to run and can be retrieved using colData().

a1

Optional numeric value used in the optimization process. If defined by the user, no optimization occurs.

a2

Optional numeric value used in the optimization process. If defined by the user, no optimization occurs.

pSamples

a permutated list of samples

Value

A list containing: D, S, pD, pS for bootstrapped data and for permuted data.